pyproject.toml
[project]
name = "swag-labs-web-tests"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pytest>=8.0",
  "pytest-xdist>=3.0",
  "python-dotenv>=1.0",
  "playwright>=1.40",
  "pytest-html>=4.0",
]
